added `ScheduledJob` decorator
