+++ package.json
@@ -27,6 +27,7 @@
     "discord.js": "^13.3.1",
     "dotenv": "^10.0.0",
     "glob": "^7.2.0",
+    "node-cron": "^3.0.0",
     "pg": "^8.7.1",
     "reflect-metadata": "^0.1.13",
     "typeorm": "^0.2.41",
@@ -37,6 +38,7 @@
   "devDependencies": {
     "@types/glob": "^7.2.0",
     "@types/node": "^16.11.13",
+    "@types/node-cron": "^3.0.1",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "eslint": "^8.4.1",
