+++ src/decorators/index.ts
@@ -3,3 +3,4 @@ export * from "./Button";
 export * from "./CommandData";
 export * from "./EventHandler";
 export * from "./AfterCommandUpdate";
+export * from "./ScheduledJob";
